# Fix close deadlock and Sub type error