@Kubuxu 2019-06-22
Would be great to have a test for #10.