@Kubuxu 2019-06-22
I've improved the test for #10